refactor(ProjectChatSessionList): inline ProjectChatMorePopup; wire interaction="active" on popover open
