--- web/src/sections/projects/ProjectChatMorePopup.tsx
@@ -1,289 +1,0 @@
-"use client";
-
-import { ChatSession } from "@/app/app/interfaces";
-import { deleteChatSession } from "@/app/app/services/lib";
-import { useProjectsContext } from "@/providers/ProjectsContext";
-import {
-  moveChatSession as moveChatSessionService,
-  removeChatSessionFromProject as removeChatSessionFromProjectService,
-} from "@/app/app/projects/projectsService";
-import { Popover, PopoverMenu } from "@opal/components";
-import { FiMoreHorizontal } from "react-icons/fi";
-import useChatSessions from "@/hooks/useChatSessions";
-import { useCallback, useState, useMemo } from "react";
-import MoveCustomAgentChatModal from "@/sections/modals/MoveCustomAgentChatModal";
-import { noProp } from "@/lib/utils";
-import { cn } from "@opal/utils";
-import ConfirmationModalLayout from "@/refresh-components/layouts/ConfirmationModalLayout";
-import { Button } from "@opal/components";
-import { PopoverSearchInput } from "@/sections/sidebar/ChatButton";
-import LineItem from "@/refresh-components/buttons/LineItem";
-import { SvgFolder, SvgFolderIn, SvgShare, SvgTrash } from "@opal/icons";
-import { DEFAULT_AGENT_ID } from "@/lib/constants";
-
-const LS_HIDE_MOVE_CUSTOM_AGENT_MODAL_KEY = "onyx:hideMoveCustomAgentModal";
-
-interface ProjectChatMorePopupProps {
-  chatSession: ChatSession;
-  projectId?: number;
-  isRenamingChat: boolean;
-  setIsRenamingChat: (value: boolean) => void;
-  showShareModal?: (chatSession: ChatSession) => void;
-  afterDelete?: () => void;
-  afterMove?: () => void;
-  afterRemoveFromProject?: () => void;
-  search?: boolean;
-  iconSize?: number;
-  isVisible?: boolean;
-}
-
-export function ProjectChatMorePopup({
-  chatSession,
-  projectId,
-  isRenamingChat: _isRenamingChat,
-  setIsRenamingChat: _setIsRenamingChat,
-  showShareModal,
-  afterDelete,
-  afterMove,
-  afterRemoveFromProject,
-  search,
-  iconSize = 16,
-  isVisible = false,
-}: ProjectChatMorePopupProps) {
-  const [popoverOpen, setPopoverOpen] = useState(false);
-  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-  const { refreshChatSessions, removeSession } = useChatSessions();
-  const { fetchProjects, projects } = useProjectsContext();
-
-  const [pendingMoveProjectId, setPendingMoveProjectId] = useState<
-    number | null
-  >(null);
-  const [showMoveCustomAgentModal, setShowMoveCustomAgentModal] =
-    useState(false);
-
-  const isChatUsingDefaultAgent = chatSession.persona_id === DEFAULT_AGENT_ID;
-
-  const [showMoveOptions, setShowMoveOptions] = useState(false);
-  const [searchTerm, setSearchTerm] = useState("");
-
-  const filteredProjects = projects.filter((project) =>
-    project.name.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  const handlePopoverOpenChange = useCallback((open: boolean) => {
-    setPopoverOpen(open);
-  }, []);
-
-  const handleConfirmDelete = useCallback(
-    async (e: React.MouseEvent<HTMLButtonElement>) => {
-      e.stopPropagation();
-      await deleteChatSession(chatSession.id);
-      removeSession(chatSession.id);
-      await refreshChatSessions();
-      await fetchProjects();
-      setIsDeleteModalOpen(false);
-      setPopoverOpen(false);
-      afterDelete?.();
-    },
-    [
-      chatSession,
-      refreshChatSessions,
-      removeSession,
-      fetchProjects,
-      afterDelete,
-    ]
-  );
-
-  const performMove = useCallback(
-    async (targetProjectId: number) => {
-      await moveChatSessionService(targetProjectId, chatSession.id);
-      await fetchProjects();
-      await refreshChatSessions();
-      setPopoverOpen(false);
-      afterMove?.();
-    },
-    [chatSession.id, fetchProjects, refreshChatSessions, afterMove]
-  );
-
-  const handleMoveChatSession = useCallback(
-    async (item: { id: number; label: string }) => {
-      const targetProjectId = item.id;
-      const hideModal =
-        typeof window !== "undefined" &&
-        window.localStorage.getItem(LS_HIDE_MOVE_CUSTOM_AGENT_MODAL_KEY) ===
-          "true";
-
-      if (!isChatUsingDefaultAgent && !hideModal) {
-        setPendingMoveProjectId(targetProjectId);
-        setShowMoveCustomAgentModal(true);
-        return;
-      }
-
-      await performMove(targetProjectId);
-    },
-    [isChatUsingDefaultAgent, performMove]
-  );
-
-  const handleRemoveChatSessionFromProject = useCallback(async () => {
-    await removeChatSessionFromProjectService(chatSession.id);
-    await fetchProjects();
-    await refreshChatSessions();
-    afterRemoveFromProject?.();
-    setPopoverOpen(false);
-  }, [
-    chatSession.id,
-    fetchProjects,
-    refreshChatSessions,
-    removeChatSessionFromProjectService,
-    afterRemoveFromProject,
-  ]);
-
-  // Build popover items similar to AppSidebar (no rename here)
-  const popoverItems = useMemo(() => {
-    if (!showMoveOptions) {
-      return [
-        showShareModal && (
-          <LineItem
-            key="share"
-            icon={SvgShare}
-            onClick={noProp(() => showShareModal(chatSession))}
-          >
-            Share
-          </LineItem>
-        ),
-        <LineItem
-          key="move"
-          icon={SvgFolderIn}
-          onClick={noProp(() => setShowMoveOptions(true))}
-        >
-          Move to Project
-        </LineItem>,
-        projectId && (
-          <LineItem
-            key="remove"
-            icon={SvgFolder}
-            onClick={noProp(() => handleRemoveChatSessionFromProject())}
-          >
-            {`Remove from ${
-              projects.find((p) => p.id === projectId)?.name ?? "Project"
-            }`}
-          </LineItem>
-        ),
-        null,
-        <LineItem
-          key="delete"
-          icon={SvgTrash}
-          onClick={noProp(() => setIsDeleteModalOpen(true))}
-          danger
-        >
-          Delete
-        </LineItem>,
-      ];
-    }
-    return [
-      <PopoverSearchInput
-        key="search"
-        setShowMoveOptions={setShowMoveOptions}
-        onSearch={setSearchTerm}
-      />,
-      ...filteredProjects
-        .filter((candidate) => candidate.id !== projectId)
-        .map((target) => (
-          <LineItem
-            key={target.id}
-            icon={SvgFolder}
-            onClick={noProp(() =>
-              handleMoveChatSession({ id: target.id, label: target.name })
-            )}
-          >
-            {target.name}
-          </LineItem>
-        )),
-    ];
-  }, [
-    showMoveOptions,
-    showShareModal,
-    projects,
-    projectId,
-    filteredProjects,
-    chatSession,
-    setShowMoveOptions,
-    setSearchTerm,
-    handleMoveChatSession,
-    handleRemoveChatSessionFromProject,
-  ]);
-
-  return (
-    <div>
-      <div className="-my-1">
-        <Popover open={popoverOpen} onOpenChange={handlePopoverOpenChange}>
-          <Popover.Trigger
-            asChild
-            onClick={(event) => {
-              event.preventDefault();
-              event.stopPropagation();
-              handlePopoverOpenChange(!popoverOpen);
-            }}
-          >
-            <div
-              className={cn(
-                "p-1 rounded-sm cursor-pointer select-none transition-opacity duration-150",
-                isVisible || popoverOpen
-                  ? "opacity-100 pointer-events-auto"
-                  : "opacity-0 pointer-events-none"
-              )}
-            >
-              <FiMoreHorizontal size={iconSize} />
-            </div>
-          </Popover.Trigger>
-          <Popover.Content
-            align="end"
-            side="right"
-            avoidCollisions
-            sideOffset={8}
-          >
-            <PopoverMenu>{popoverItems}</PopoverMenu>
-          </Popover.Content>
-        </Popover>
-      </div>
-      {isDeleteModalOpen && (
-        <ConfirmationModalLayout
-          title="Delete Chat"
-          icon={SvgTrash}
-          onClose={() => setIsDeleteModalOpen(false)}
-          submit={
-            <Button variant="danger" onClick={handleConfirmDelete}>
-              Delete
-            </Button>
-          }
-        >
-          Are you sure you want to delete this chat? This action cannot be
-          undone.
-        </ConfirmationModalLayout>
-      )}
-
-      {showMoveCustomAgentModal && (
-        <MoveCustomAgentChatModal
-          onCancel={() => {
-            setShowMoveCustomAgentModal(false);
-            setPendingMoveProjectId(null);
-          }}
-          onConfirm={async (doNotShowAgain: boolean) => {
-            if (doNotShowAgain && typeof window !== "undefined") {
-              window.localStorage.setItem(
-                LS_HIDE_MOVE_CUSTOM_AGENT_MODAL_KEY,
-                "true"
-              );
-            }
-            const target = pendingMoveProjectId;
-            setShowMoveCustomAgentModal(false);
-            setPendingMoveProjectId(null);
-            if (target != null) {
-              await performMove(target);
-            }
-          }}
-        />
-      )}
-    </div>
-  );
-}
